Comments

+++ ft_errors.c
@@ -12,6 +12,19 @@
 
 #include "pipex.h"
 
+/** free_arguments_cmd:
+ * This function desalocate the arguments of the command.
+ */
+void	free_arguments_cmd(t_inf *info)
+{
+	int	i;
+
+	i = -1;
+	while(info->args_cmd[++i] !=  NULL)
+		free(info->args_cmd[i]);
+	free(info->args_cmd);
+}
+
 /** free_memory:
  * This function is called when something went wrong.
  * Release all information and exit the program whith Failure code(1).
@@ -23,12 +36,16 @@ void	free_memory(char *str1,char *str2, char *str3, t_inf *info)
 	ft_putstr_fd(str2, 2);
 	ft_putendl_fd(str3, 2);
 
-	if (ft_strcmp(info->argv[1], ".here\\_doc") == 0)
-		unlink(info->argv[1]);
-	if (info->pid != NULL)
-		free(info->pid);
-	if (info->fds != NULL)
-		free(info->fds);
+	if (info != NULL){
+		if (ft_strcmp(info->argv[1], ".here\\_doc") == 0)
+			unlink(info->argv[1]);
+		if (info->pid != NULL)
+			free(info->pid);
+		if (info->fds != NULL)
+			free(info->fds);
+		if (info->args_cmd != NULL)
+			free_arguments_cmd(info);
+	}
 	exit(EXIT_FAILURE);
 }
